[RESTRICT AUTOMERGE] Quit cec-client process correctly.
Currently, the test kill cec-client process forcebly.
Then, CEC Port is not closed correctly.
and when the test open CEC Port again
as the different device type,
Pysical Address cannot be gotten.
the test should quit cec-client process correctly.

It need to send "q" command
and receive the messsage "communication thread ended"

Bug: 304810851
Test: atest CtsHdmiCecHostTestCases
(cherry picked from commit 366341cdb943efee7c7225e9bdb8d0fab7cab1dc)

+++ hostsidetests/hdmicec/src/android/hdmicec/cts/targetprep/CecPortDiscoverer.java
@@ -17,6 +17,7 @@
 package android.hdmicec.cts.targetprep;
 
 import android.hdmicec.cts.BaseHdmiCecCtsTest;
+import android.hdmicec.cts.CecClientMessage;
 import android.hdmicec.cts.CecMessage;
 import android.hdmicec.cts.HdmiCecClientWrapper;
 import android.hdmicec.cts.HdmiCecConstants;
@@ -35,6 +36,7 @@ import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.concurrent.TimeUnit;
@@ -139,13 +141,17 @@ public class CecPortDiscoverer extends BaseTargetPreparer {
                         break;
                     }
                     mCecClient = RunUtil.getDefault().runCmdInBackground(launchCommand);
-                    try (BufferedReader inputConsole =
+                    BufferedReader inputConsole =
                             new BufferedReader(
-                                    new InputStreamReader(mCecClient.getInputStream()))) {
-
+                                    new InputStreamReader(mCecClient.getInputStream()));
+                    BufferedWriter outputConsole =
+                            new BufferedWriter(
+                                    new OutputStreamWriter(mCecClient.getOutputStream()));
+                    try {
                         /* Wait for the client to become ready */
                         if (cecClientWrapper.checkConsoleOutput(
-                                "waiting for input", TIMEOUT_MILLIS, inputConsole)) {
+                                CecClientMessage.CLIENT_CONSOLE_READY.toString(),
+                                TIMEOUT_MILLIS, inputConsole)) {
 
                             device.executeShellCommand(sendVendorCommand.toString());
                             if (cecClientWrapper.checkConsoleOutput(
@@ -170,8 +176,26 @@ public class CecPortDiscoverer extends BaseTargetPreparer {
                         }
                     } finally {
                         /* Kill the unwanted cec-client process. */
-                        Process killProcess = mCecClient.destroyForcibly();
-                        killProcess.waitFor(60, TimeUnit.SECONDS);
+                        boolean processQuit = false;
+                        outputConsole.write(CecClientMessage.QUIT_CLIENT.toString());
+                        outputConsole.flush();
+                        if (cecClientWrapper.checkConsoleOutput(
+                                CecClientMessage.CLIENT_CONSOLE_END.toString(),
+                                TIMEOUT_MILLIS, inputConsole)) {
+                            outputConsole.close();
+                            inputConsole.close();
+                            if (mCecClient.waitFor(10, TimeUnit.SECONDS)) {
+                                /* The cec-client process is quit */
+                                processQuit = true;
+                            }
+                        }
+                        if (!processQuit) {
+                            /* Use destryForcibly if the cec-client process is not dead
+                             * in spite of the quit above.
+                             */
+                            Process killProcess = mCecClient.destroyForcibly();
+                            killProcess.waitFor(60, TimeUnit.SECONDS);
+                        }
                     }
                 } while (portBeingRetried);
                 launchCommand.remove(port);
